+++ src/Commands/Revert.php
@@ -37,21 +37,12 @@ class Revert extends Command
      */
     public function handle()
     {
-        $commitID = $this->ask('commit ID [and press enter]');
-        $branch = $this->ask('branch name [and press enter]');
+        $tagname = $this->ask('Tag version [and press enter]');
+        $this->alert('~# init basher:tag');
 
-        $this->alert('~# init basher:revert');
+        exec('git tag -a ' . $tagname . ' -m "Versión ' . $tagname . '"');
+        exec('git push origin '.$tagname);
 
-        exec('git reset ' . $commitID . ' --hard');
-        if (empty($branch) || $branch == '') {
-            $this->info('~# push --force to branch master');
-            exec('git push origin master --force');
-        } else {
-            $this->info('~# push --force to branch ' . $branch);
-            exec('git push origin ' . $branch . '  --force');
-        }
-        exec('git status');
-
-        $this->alert('~# end basher:revert');
+        $this->alert('~# end basher:tag');
     }
 }